requirements.txt
blinker==1.6.2
boto3==1.28.49
botocore==1.31.49
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
docopt==0.6.2
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
pipreqs==0.4.13
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
yarg==0.1.9
